Refactored datacommons api and db directory structures
