+++ datacommons/datacommons-api/src/datacommons/api/endpoints/routers/node_router.py
@@ -1,0 +1,39 @@
+import traceback
+from fastapi import APIRouter, Depends
+from datacommons.schema.models.jsonld import JSONLDDocument
+from datacommons.api.services.graph_service import GraphService
+from datacommons.api.core.logging import get_logger
+from datacommons.api.endpoints.responses import UpdateResponse
+from datacommons.api.endpoints.dependencies import with_graph_service
+
+logger = get_logger(__name__)
+
+router = APIRouter()
+
+# JSON-LD endpoint
+@router.get("/nodes/", response_model=JSONLDDocument, response_model_exclude_none=True)
+def get_nodes(
+  limit: int = 100,
+  type: str = None,
+  graph_service: GraphService = Depends(with_graph_service)
+) -> JSONLDDocument:
+  """
+  Get nodes with their edges
+  """
+  # Get nodes with their edges
+  response_document = graph_service.get_graph_nodes(limit=limit, type_filter=type)
+  return response_document
+
+@router.post("/nodes/", response_model=UpdateResponse, response_model_exclude_none=True)
+def insert_nodes(
+  jsonld: JSONLDDocument,
+  graph_service: GraphService = Depends(with_graph_service)
+) -> UpdateResponse:
+  """Insert a JSON-LD document into the database"""
+  try:
+    graph_service.insert_graph_nodes(jsonld)
+    return UpdateResponse(success=True, message=f"Inserted {len(jsonld.graph)} nodes successfully")
+  except Exception as e:
+    logger.error(f"Error inserting nodes: {e}\n{traceback.format_exc()}")
+    return UpdateResponse(success=False, message=str(e))
+
