+++ datacommons/datacommons-db/src/datacommons/db/session.py
@@ -1,9 +1,12 @@
 from sqlalchemy import create_engine, inspect, Engine
 from sqlalchemy.orm import sessionmaker, Session
-from .base import Base
+from datacommons.db.models.base import Base
+from datacommons.db.models.node import NodeModel
+from datacommons.db.models.edge import EdgeModel
+from datacommons.db.models.observation import ObservationModel
 import logging
 
-def get_spanner_engine(project_id: str, instance_id: str, database_name: str) -> Engine:
+def get_engine(project_id: str, instance_id: str, database_name: str) -> Engine:
   """Create and return a SQLAlchemy engine for Cloud Spanner.
   
   Args:
@@ -16,10 +19,9 @@ def get_spanner_engine(project_id: str, instance_id: str, database_name: str) ->
   """
   return create_engine(
     f"spanner+spanner:///projects/{project_id}/instances/{instance_id}/databases/{database_name}",
-    # connect_args={'client': spanner.Client(project=project_id)}
   )
 
-def get_spanner_session(project_id: str, instance_id: str, database_name: str) -> Session:
+def get_session(project_id: str, instance_id: str, database_name: str) -> Session:
   """Create and return a SQLAlchemy session for Cloud Spanner.
   
   Args:
@@ -30,7 +32,7 @@ def get_spanner_session(project_id: str, instance_id: str, database_name: str) -
   Returns:
     SQLAlchemy session configured for Cloud Spanner
   """
-  engine = get_spanner_engine(project_id, instance_id, database_name)
+  engine = get_engine(project_id, instance_id, database_name)
   Session = sessionmaker(bind=engine)
   return Session()
 
@@ -42,7 +44,7 @@ def initialize_db(project_id: str, instance_id: str, database_name: str):
     instance_id: Cloud Spanner instance ID
     database_name: Cloud Spanner database name
   """
-  engine = get_spanner_engine(project_id, instance_id, database_name)
+  engine = get_engine(project_id, instance_id, database_name)
   
   # Check if database is empty by inspecting existing tables
   inspector = inspect(engine)
